6419 | 18 files | M DumpSource2/schemas/client/CDOTA_Ability_VoidSpirit_ResonantPulse.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Broodmother_Spiders_Milk.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Invoker_IceWall_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Invoker_IceWall_VectorTarget_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_StormSpirit_Galvanized.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Visage_SummonFamiliars_In_Formation.h,  M DumpSource2/schemas/client/C_DOTA_Ability_OgreBruiser_OgreSmash.h,  M DumpSource2/schemas/server/CDOTA_Ability_OgreBruiser_OgreSmash.h,  M DumpSource2/schemas/server/CDOTA_Ability_VoidSpirit_ResonantPulse.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Broodmother_Spiders_Milk.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Invoker_IceWall_Thinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Invoker_IceWall_VectorTarget_Thinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_StormSpirit_Galvanized.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Visage_Summ
https://steamdb.info/patchnotes/18585611/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Broodmother_Spiders_Milk.h
@@ -2,6 +2,6 @@ class CDOTA_Modifier_Broodmother_Spiders_Milk : public CDOTA_Buff
 {
 	float32 kill_heal_heroes;
 	int32 kill_creeps_penalty;
-	int32 kill_heal_aoe;
+	float32 kill_heal_aoe;
 	float32 kill_heal_duration;
 };
