+++ DumpSource2/schemas/client/CDOTA_Modifier_Invoker_IceWall_Thinker.h
@@ -3,8 +3,8 @@ class CDOTA_Modifier_Invoker_IceWall_Thinker : public CDOTA_Buff
 	float32 damage_per_second;
 	Vector m_vWallDirection;
 	Vector m_vWallStart;
-	int32 wall_total_length;
-	int32 wall_width;
+	float32 wall_total_length;
+	float32 wall_width;
 	float32 slow_duration;
 	int32 slow;
 	float32 root_damage;
