+++ DumpSource2/schemas/server/CDOTA_Ability_OgreBruiser_OgreSmash.h
@@ -1,8 +1,8 @@
 class CDOTA_Ability_OgreBruiser_OgreSmash : public CDOTABaseAbility
 {
-	int32 radius;
+	float32 radius;
 	float32 hero_stun_duration;
-	int32 damage;
+	float32 damage;
 	float32 damage_pct;
 	GameTime_t flCooldown;
 };
