+++ DumpSource2/schemas/server/CDOTA_Ability_VoidSpirit_ResonantPulse.h
@@ -3,5 +3,5 @@ class CDOTA_Ability_VoidSpirit_ResonantPulse : public CDOTABaseAbility
 	bool m_bGrantedScepterCharges;
 	float32 buff_duration;
 	float32 base_absorb_amount;
-	int32 absorb_per_hero_hit;
+	float32 absorb_per_hero_hit;
 };
