+++ DumpSource2/schemas/server/CDOTA_Modifier_Visage_SummonFamiliars_In_Formation.h
@@ -6,9 +6,9 @@ class CDOTA_Modifier_Visage_SummonFamiliars_In_Formation : public CDOTA_Buff
 	int32 back_distance;
 	int32 side_distance;
 	Vector m_vecDesiredPosition;
-	int32 max_distance;
-	int32 return_distance;
-	int32 attack_range_buffer;
+	float32 max_distance;
+	float32 return_distance;
+	float32 attack_range_buffer;
 	float32 recall_duration;
 	bool m_bTeleporting;
 	int32 familiar_attack_range;
